requirements.txt
gradio==6.2.0
jinja2==3.1.6
python-dotenv==1.2.1
httpx==0.28.1

langchain==1.0.5
langchain-core==1.0.4
langchain-openai==1.0.2

pydantic==2.12.4
pydantic-core==2.41.5
openai==2.7.1